make payment route

+++ server/internal/models/payments/payment.go
@@ -2,7 +2,7 @@ package models
 
 type Payment struct {
 	Id      int
-  TotalValue   float64
-  PayedValue   float64
+  TotalValue   int64
+  PayedValue   int64
 }
